Added project to create binary with bindings. This can be used to create docker image. Refactored integration test code.
