+++ build.gradle
@@ -181,7 +181,6 @@ project('segmentstore:storage:impl') {
         testCompile group: 'org.apache.hadoop', name: 'hadoop-minicluster', version: hadoopVersion, withoutLog4j
         testCompile group: 'ch.qos.logback', name: 'logback-classic', version: qosLogbackVersion
 
-        //For Extended S3
         compile project(':common')
         compile project(':segmentstore:storage')
         compile project(':shared:metrics')
@@ -193,29 +192,31 @@ project('segmentstore:storage:impl') {
 
 project('segmentstore:storage:bindings') {
     dependencies {
-        compile group: 'org.apache.bookkeeper', name: 'bookkeeper-server', version: bookKeeperVersion, withoutLog4j
+        dependencies {
+            compile group: 'org.apache.bookkeeper', name: 'bookkeeper-server', version: bookKeeperVersion, withoutLog4j
 
-        compile group: 'org.rocksdb', name: 'rocksdbjni', version: rocksdbjniVersion
+            compile group: 'org.rocksdb', name: 'rocksdbjni', version: rocksdbjniVersion
 
-        // https://mvnrepository.com/artifact/org.apache.hadoop/hadoop-common
-        compile group: 'org.apache.hadoop', name: 'hadoop-common', version:hadoopVersion, withoutLog4j
+            // https://mvnrepository.com/artifact/org.apache.hadoop/hadoop-common
+            compile group: 'org.apache.hadoop', name: 'hadoop-common', version:hadoopVersion, withoutLog4j
 
-        // https://mvnrepository.com/artifact/org.apache.hadoop/hadoop-common
-        compile group: 'org.apache.hadoop', name: 'hadoop-hdfs', version: hadoopVersion, withoutLog4j
-        testCompile group: 'org.apache.hadoop', name: 'hadoop-minicluster', version: hadoopVersion, withoutLog4j
-        testCompile group: 'ch.qos.logback', name: 'logback-classic', version: qosLogbackVersion
+            // https://mvnrepository.com/artifact/org.apache.hadoop/hadoop-common
+            compile group: 'org.apache.hadoop', name: 'hadoop-hdfs', version: hadoopVersion, withoutLog4j
+            testCompile group: 'org.apache.hadoop', name: 'hadoop-minicluster', version: hadoopVersion, withoutLog4j
+            testCompile group: 'ch.qos.logback', name: 'logback-classic', version: qosLogbackVersion
 
-        //For Extended S3
-        compile "com.emc.ecs:object-client:3.0.0"
-        testCompile group: 'org.gaul', name: 's3proxy', version: '1.5.2'
-        testCompile group: 'commons-httpclient', name: 'commons-httpclient', version: '3.1'
+            //For Extended S3
+            compile "com.emc.ecs:object-client:3.0.0"
+            testCompile group: 'org.gaul', name: 's3proxy', version: '1.5.2'
+            testCompile group: 'commons-httpclient', name: 'commons-httpclient', version: '3.1'
 
-        compile project(':common')
-        compile project(':segmentstore:storage')
-        compile project(':shared:metrics')
-        testCompile project(':test:testcommon')
-        testCompile project(path:':segmentstore:storage', configuration:'testRuntime')
-        testCompile project(path:':segmentstore:storage:impl', configuration:'testRuntime')
+            compile project(':common')
+            compile project(':segmentstore:storage')
+            compile project(':shared:metrics')
+            testCompile project(':test:testcommon')
+            testCompile project(path:':segmentstore:storage', configuration:'testRuntime')
+        }
+
     }
 }
 
@@ -258,6 +259,62 @@ project('segmentstore:server:host') {
         testCompile group: 'org.apache.hadoop', name: 'hadoop-minicluster', version: hadoopVersion, withoutLog4j
         testCompile project(path:':segmentstore:storage:impl', configuration:'testRuntime')
         testCompile group: 'commons-httpclient', name: 'commons-httpclient', version: '3.1'
+    }
+
+    task createAppWithGCLogging(type: CreateStartScripts) {
+        applicationName = "pravega-segmentstore-withGCLogging"
+        mainClassName = "io.pravega.controller.server.Main"
+        defaultJvmOpts = ["-server", "-Xms512m", "-XX:+HeapDumpOnOutOfMemoryError",
+                          "-XX:+PrintGCDetails", "-XX:+PrintGCDateStamps",
+                          "-Xloggc:PRAVEGA_APP_HOME/logs/gc.log", "-XX:+UseGCLogFileRotation",
+                          "-XX:NumberOfGCLogFiles=2", "-XX:GCLogFileSize=64m",
+                          "-Dlog.dir=PRAVEGA_APP_HOME/logs",
+                          "-Dlog.name=segmentstore",
+                          "-Dpravega.configurationFile=PRAVEGA_APP_HOME/conf/config.properties",
+                          "-Dlogback.configurationFile=PRAVEGA_APP_HOME/conf/logback.xml"]
+        classpath = startScripts.classpath
+        outputDir = startScripts.outputDir
+    }
+
+    applicationDistribution.into("bin") {
+        from(createAppWithGCLogging)
+    }
+}
+
+
+project('segmentstore:server:hostwithbindings') {
+    apply plugin: 'application'
+    applicationName = "pravega-segmentstore"
+    mainClassName = "io.pravega.segmentstore.server.host.ServiceStarter"
+
+    applicationDefaultJvmArgs = ["-server", "-Xms512m", "-XX:+HeapDumpOnOutOfMemoryError",
+                                 "-Dlog.dir=PRAVEGA_APP_HOME/logs",
+                                 "-Dlog.name=segmentstore",
+                                 "-Dpravega.configurationFile=PRAVEGA_APP_HOME/conf/config.properties",
+                                 "-Dlogback.configurationFile=PRAVEGA_APP_HOME/conf/logback.xml"]
+
+    applicationDistribution.from('src/config') {
+        into "conf"
+    }
+
+    dependencies {
+        compile project(':common')
+        compile project(':segmentstore:contracts')
+        compile project(':client')
+        compile project(':segmentstore:storage')
+        compile project(':segmentstore:storage:impl')
+        compile project(':segmentstore:storage:bindings')
+        compile project(':segmentstore:server')
+        compile group: 'ch.qos.logback', name: 'logback-classic', version: qosLogbackVersion
+        testCompile project(':test:testcommon')
+        testCompile group: 'org.apache.curator', name: 'curator-test', version: apacheCuratorVersion
+        testCompile project(path:':segmentstore:server', configuration:'testRuntime')
+        testCompile group: 'org.apache.hadoop', name: 'hadoop-minicluster', version: hadoopVersion, withoutLog4j
+        testCompile project(path:':segmentstore:storage:impl', configuration:'testRuntime')
+        testCompile project(path:':segmentstore:storage:bindings', configuration:'testRuntime')
+        testCompile group: 'commons-httpclient', name: 'commons-httpclient', version: '3.1'
+        testCompile project(path:':segmentstore:server', configuration:'testRuntime')
+        testCompile project(path:':segmentstore:server:host', configuration:'testRuntime')
     }
 
     task createAppWithGCLogging(type: CreateStartScripts) {
