+++ segmentstore/server/hostwithbindings/src/test/java/io/pravega/segmentstore/host/bindings/FileSystemIntegrationTest.java
@@ -7,17 +7,18 @@
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.storage.bindings.filesystem;
+package io.pravega.segmentstore.host.bindings;
 
 import io.pravega.common.io.FileHelpers;
+import io.pravega.common.util.ServiceBuilderConfig;
+import io.pravega.segmentstore.server.host.BookKeeperRunner;
 import io.pravega.segmentstore.server.store.ServiceBuilder;
-import io.pravega.segmentstore.server.store.ServiceBuilderConfig;
 import io.pravega.segmentstore.server.store.StreamSegmentStoreTestBase;
 import io.pravega.segmentstore.storage.StorageFactory;
-import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperConfig;
-import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperLogFactory;
 import io.pravega.segmentstore.storage.bindings.filesystem.FileSystemStorageConfig;
 import io.pravega.segmentstore.storage.bindings.filesystem.FileSystemStorageFactory;
+import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperConfig;
+import io.pravega.segmentstore.storage.impl.bookkeeper.BookKeeperLogFactory;
 import io.pravega.segmentstore.storage.impl.rocksdb.RocksDBCacheFactory;
 import io.pravega.segmentstore.storage.impl.rocksdb.RocksDBConfig;
 import java.io.File;
