+++ segmentstore/server/hostwithbindings/src/test/java/io/pravega/segmentstore/host/bindings/S3ClientWrapper.java
@@ -7,7 +7,7 @@
  *
  * http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.pravega.segmentstore.server.host;
+package io.pravega.segmentstore.host.bindings;
 
 import com.emc.object.Range;
 import com.emc.object.s3.S3Config;
@@ -25,7 +25,7 @@ import com.emc.object.s3.request.CopyPartRequest;
 import com.emc.object.s3.request.DeleteObjectsRequest;
 import com.emc.object.s3.request.PutObjectRequest;
 import com.emc.object.s3.request.SetObjectAclRequest;
-import io.pravega.segmentstore.storage.impl.extendeds3.S3ImplBase;
+import io.pravega.segmentstore.storage.bindings.extendeds3.S3ImplBase;
 import java.io.InputStream;
 import lombok.Synchronized;
