+++ settings.gradle
@@ -22,6 +22,7 @@ include 'common',
         'segmentstore:storage:bindings',
         'segmentstore:server',
         'segmentstore:server:host',
+        'segmentstore:server:hostwithbindings',
         'controller',
         'standalone',
         'test:testcommon',
